Change: No gbkviz webapp launch test on MacOS

+++ tests/test_gbkviz_webapp.py
@@ -1,4 +1,5 @@
 import multiprocessing
+import platform
 import sys
 import time
 
@@ -8,10 +9,13 @@ from streamlit.cli import main as stcli_main
 
 def test_gbkviz_webapp_launch():
     """test gbkviz webapp launch propery"""
-    sys.argv = ["streamlit", "run", gbkviz_webapp.__file__]
-    p = multiprocessing.Process(target=stcli_main, name="st")
-    p.start()
-    time.sleep(10)
-    p.terminate()
-    p.join()
-    assert p.exitcode == 0
+    # TODO: Test only Linux now. This test method is not applicable to MacOS.
+    #       Test MacOS as well in future.
+    if platform.system() == "Linux":
+        sys.argv = ["streamlit", "run", gbkviz_webapp.__file__]
+        p = multiprocessing.Process(target=stcli_main, name="st")
+        p.start()
+        time.sleep(10)
+        p.terminate()
+        p.join()
+        assert p.exitcode == 0
